Initial Commit
TunnelTest was used for debugging of the hardware, software, and physical construction of the tunnel. TunnelMatlabCode ran the Wind Tunnel for a planned amount of time. Readings could be made to assess the accuracy of the predicted output of the device. TunnelLoop would be used for regular function of the Wind Tunnel and allow the user to produce known velocity profiles on their test subjects.